Move DAP hacks to dap-hacks.js, require() in both bundles.
